Update on progress
I went ahead and created all the components that were needed. Then I started on the Main.java and created an ArrayList of the piggybank so that I can add all the new coin values, then I printed out each of their values, created a for loop that says if the value(number) is less than the piggybank then add it onto the piggybank, and then I printed the total of the piggybank. Then I created the Coins.java and created a numAdded as a number to create the functions of getNumAdded()  and getTotal(). Now working on each of the coin values.